Добавлены тесты для PetController

+++ ClinicService/Controllers/PetController.cs
@@ -2,6 +2,7 @@
 using ClinicService.Models;
 using ClinicService.Services;
 using Microsoft.AspNetCore.Mvc;
+using Swashbuckle.AspNetCore.Annotations;
 
 namespace ClinicService.Controllers
 {
@@ -16,10 +17,18 @@ namespace ClinicService.Controllers
             _petRepository = petRepository;
         }
 
+        public PetController()
+        {
+        }
 
         [HttpPost("create")]
+        [SwaggerOperation(OperationId = "PetCreate")]
         public ActionResult<int> Create([FromBody] CreatePetRequest createRequest)
         {
+            if (createRequest.Age > 25 || createRequest.Name.Length < 3)
+            {
+                return BadRequest(0);
+            }
             int res = _petRepository.Create(new Pet
             {
                 ClientId = createRequest.ClientId,
@@ -31,8 +40,13 @@ namespace ClinicService.Controllers
         }
 
         [HttpPut("update")]
+        [SwaggerOperation(OperationId = "PetUpdate")]
         public ActionResult<int> Update([FromBody] UpdatePetRequest updateRequest)
         {
+            if (updateRequest.Age > 25 || updateRequest.Name.Length < 3)
+            {
+                return BadRequest(0);
+            }
             int res = _petRepository.Update(new Pet
             {
                 PetId = updateRequest.PetId,
@@ -45,21 +59,32 @@ namespace ClinicService.Controllers
         }
 
         [HttpDelete("delete")]
+        [SwaggerOperation(OperationId = "PetDelete")]
         public ActionResult<int> Delete(int petId)
         {
+            if(petId <= 0)
+            {
+                return BadRequest(0);
+            }
             int res = _petRepository.Delete(petId);
             return Ok(res);
         }
 
         [HttpGet("get-all")]
+        [SwaggerOperation(OperationId = "PetGetAll")]
         public ActionResult<List<Pet>> GetAll()
         {
             return Ok(_petRepository.GetAll());
         }
 
         [HttpGet("get-by-id")]
+        [SwaggerOperation(OperationId = "PetGetById")]
         public ActionResult<Pet> GetById(int petId)
         {
+            if(petId <= 0)
+            {
+                return BadRequest(null);
+            }
             return Ok(_petRepository.GetById(petId));
         }
     }
